Fix crash on missing database
DeleteResourceIdsInBatch will return STATUS_OK if LazyOpen(false) can't
find a database. In this case db_ would be null and we shouldn't call
WriteBatch() to avoid crashing.

Bug: 931160
Reviewed-on: https://chromium-review.googlesource.com/c/1474768
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#633646}(cherry picked from commit 7f5bb07e152ef23a31577fd02f3439c597fe1977)
Reviewed-on: https://chromium-review.googlesource.com/c/1479997
Cr-Commit-Position: refs/branch-heads/3683@{#556}
Cr-Branched-From: e51029943e0a38dd794b73caaf6373d5496ae783-refs/heads/master@{#625896}

+++ content/browser/service_worker/service_worker_database.cc
@@ -1192,8 +1192,12 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::GetPurgeableResourceIds(
 
 ServiceWorkerDatabase::Status ServiceWorkerDatabase::ClearPurgeableResourceIds(
     const std::set<int64_t>& ids) {
+  Status status = LazyOpen(false);
+  if (IsNewOrNonexistentDatabase(status))
+    return STATUS_OK;
+
   leveldb::WriteBatch batch;
-  Status status = DeleteResourceIdsInBatch(
+  status = DeleteResourceIdsInBatch(
       service_worker_internals::kPurgeableResIdKeyPrefix, ids, &batch);
   if (status != STATUS_OK)
     return status;
@@ -1203,8 +1207,12 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::ClearPurgeableResourceIds(
 ServiceWorkerDatabase::Status
 ServiceWorkerDatabase::PurgeUncommittedResourceIds(
     const std::set<int64_t>& ids) {
+  Status status = LazyOpen(false);
+  if (IsNewOrNonexistentDatabase(status))
+    return STATUS_OK;
+
   leveldb::WriteBatch batch;
-  Status status = DeleteResourceIdsInBatch(
+  status = DeleteResourceIdsInBatch(
       service_worker_internals::kUncommittedResIdKeyPrefix, ids, &batch);
   if (status != STATUS_OK)
     return status;
